feat: db features, server actions set up and save table topic to DB
