+++ app/server/actions.ts
@@ -1,0 +1,9 @@
+"use server";
+
+import { getRandomTableTopic } from "./db/queries";
+
+export async function getTableTopic() {
+  const topic = await getRandomTableTopic();
+
+  return topic;
+}
