+++ app/server/db/topics.json
@@ -1,0 +1,53 @@
+[
+  "Describe a moment when you stepped out of your comfort zone.",
+  "What’s the best advice you’ve ever received?",
+  "Share a time you had to make a tough decision.",
+  "What is the most adventurous thing you’ve ever done?",
+  "Describe a time when you overcame a fear.",
+  "What is the most valuable lesson you learned from a failure?",
+  "Who has been the biggest influence in your life, and why?",
+  "If you could relive one day of your life, what would it be?",
+  "If you could live anywhere in the world, where would it be and why?",
+  "If you could have dinner with any historical figure, who would it be?",
+  "What would you do if you had a million dollars but could only spend it on others?",
+  "If you could switch lives with anyone for a day, who would it be?",
+  "If you had a superpower, what would it be and why?",
+  "What would you do if you had one week to live?",
+  "If you could solve one problem in the world, what would it be?",
+  "Invent a new holiday and describe how it’s celebrated.",
+  "Describe a world where everyone has to communicate through dance.",
+  "Create a story about finding a treasure in your backyard.",
+  "What would you do if you woke up 100 years in the past?",
+  "Imagine you’ve discovered a new planet—what’s it like?",
+  "What does success mean to you?",
+  "Can money buy happiness?",
+  "Is it better to be kind or to be right?",
+  "What does it mean to live a meaningful life?",
+  "How do you define true friendship?",
+  "Do you believe in fate, or do we create our own destiny?",
+  "What role does failure play in achieving success?",
+  "What’s the funniest thing that ever happened to you?",
+  "If you could only eat one food for the rest of your life, what would it be?",
+  "What’s your favorite holiday tradition?",
+  "What’s the most unusual talent you have?",
+  "If you could be any animal for a day, which one would you choose?",
+  "What’s the most bizarre dream you’ve ever had?",
+  "Describe your perfect weekend.",
+  "What’s your take on the rise of AI in everyday life?",
+  "Should humans colonize other planets?",
+  "What’s your opinion on work-from-home culture?",
+  "What recent invention has impressed you the most?",
+  "How should society handle misinformation on social media?",
+  "What’s a recent trend that you just don’t understand?",
+  "What’s your favorite book or movie, and why?",
+  "What’s your favorite childhood memory?",
+  "If you could relive any vacation, which one would it be?",
+  "What’s your favorite season of the year and why?",
+  "Describe your dream job.",
+  "What’s your favorite hobby or way to spend free time?",
+  "What’s one thing you love about your culture?",
+  "What’s a country you’ve always wanted to visit, and why?",
+  "What do you think makes a city or town great to live in?",
+  "Describe a festival or tradition you’d like to experience.",
+  "What’s a dish from another country you’d love to try?"
+]
