+++ app/topic/create/_components/table-topics-recorder.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { getTableTopic } from "@/app/server/actions";
 import { Button } from "@/components/ui/button";
 import {
   Card,
@@ -8,125 +9,16 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
+import Spinner from "@/components/ui/spinner";
 import { cn } from "@/lib/utils";
 import { Download, Save, Trash2 } from "lucide-react";
-import { useEffect, useRef, useState } from "react";
-
-const useMediaRecorder = () => {
-  const [isRecording, setIsRecording] = useState(false);
-  const [recordedBlob, setRecordedBlob] = useState<Blob | null>(null);
-  const [recordedVideoURL, setRecordedVideoURL] = useState<string | null>(null);
-  const [timeElapsed, setTimeElapsed] = useState(0);
-
-  const mediaRecorderRef = useRef<MediaRecorder | null>(null);
-  const streamRef = useRef<MediaStream | null>(null);
-  const videoElementRef = useRef<HTMLVideoElement | null>(null);
-
-  const setVideoToStream = async () => {
-    try {
-      const stream = await navigator.mediaDevices.getUserMedia({
-        video: true,
-        audio: true,
-      });
-      streamRef.current = stream;
-
-      if (!videoElementRef.current)
-        throw new Error("Unable to set stream to video");
-
-      videoElementRef.current.srcObject = streamRef.current;
-    } catch (error) {
-      const message =
-        error instanceof Error
-          ? error.message
-          : "Failed to access camera/microphone";
-      // TODO: Add toast or alert component to show error to user and re request access to media stream
-      console.error("Error accessing media devices:", message);
-    }
-  };
-
-  const startRecording = async () => {
-    if (!streamRef.current) await setVideoToStream();
-    if (!streamRef.current) throw new Error("Unable to set media stream");
-    mediaRecorderRef.current = new MediaRecorder(streamRef.current);
-    mediaRecorderRef.current.ondataavailable = (event) => {
-      if (event.data.size > 0) {
-        setRecordedBlob(event.data);
-        setRecordedVideoURL(URL.createObjectURL(event.data));
-      }
-    };
-    mediaRecorderRef.current.start();
-    setIsRecording(true);
-  };
-
-  const stopRecording = () => {
-    if (!mediaRecorderRef.current) return;
-
-    mediaRecorderRef.current.stop();
-    if (videoElementRef.current) videoElementRef.current.srcObject = null;
-    setIsRecording(false);
-  };
-
-  const resetRecording = () => {
-    setIsRecording(false);
-    setRecordedBlob(null);
-    setRecordedVideoURL(null);
-    setVideoToStream();
-    if (recordedVideoURL) URL.revokeObjectURL(recordedVideoURL);
-    if (!streamRef.current) return;
-    streamRef.current.getTracks().forEach((track) => track.stop());
-    streamRef.current = null;
-  };
-
-  useEffect(() => {
-    let interval: NodeJS.Timeout | null = null;
-
-    if (isRecording) {
-      interval = setInterval(() => {
-        setTimeElapsed((prev) => prev + 1);
-      }, 1000);
-    } else {
-      setTimeElapsed(0);
-    }
-
-    return () => {
-      if (interval) clearInterval(interval);
-    };
-  }, [isRecording]);
-
-  useEffect(() => {
-    void setVideoToStream();
-
-    return () => {
-      if (recordedVideoURL) URL.revokeObjectURL(recordedVideoURL);
-    };
-  }, []);
-
-  return {
-    isRecording,
-    recordedBlob,
-    recordedVideoURL,
-    timeElapsed,
-    videoElementRef,
-    startRecording,
-    stopRecording,
-    resetRecording,
-  };
-};
-
-// Mock function to generate a table topic
-const generateTopic = () => {
-  const topics = [
-    "If you could have dinner with any historical figure, who would it be and why?",
-    "Describe a time when you had to think on your feet.",
-    "What's the most valuable lesson you've learned in life so far?",
-    "If you could instantly become an expert in one subject, what would it be?",
-    "Describe your perfect day from start to finish.",
-  ];
-  return topics[Math.floor(Math.random() * topics.length)];
-};
+import { useState, useTransition } from "react";
+import { useMediaRecorder } from "../_hooks";
 
 export default function TableTopicsRecorder() {
   const [currentTopic, setCurrentTopic] = useState<string | null>(null);
+  const [isPending, startTransition] = useTransition();
+
   const {
     isRecording,
     recordedVideoURL,
@@ -138,8 +30,11 @@ export default function TableTopicsRecorder() {
   } = useMediaRecorder();
 
   const handleGenerateTopic = () => {
-    setCurrentTopic(generateTopic());
-    startRecording();
+    startTransition(async () => {
+      const { topic } = await getTableTopic();
+      setCurrentTopic(topic);
+      startRecording();
+    });
   };
 
   const handleSaveRecording = () => {
@@ -181,10 +76,13 @@ export default function TableTopicsRecorder() {
     <div className="container mx-auto p-4">
       <Card className="w-full max-w-3xl mx-auto">
         <CardHeader>
-          <CardTitle>Table Topics Recorder</CardTitle>
+          <CardTitle>
+            <h1 className="text-2xl md:text-3xl">Table topics recorder</h1>
+          </CardTitle>
         </CardHeader>
         <CardContent className="space-y-4">
           <Button onClick={handleGenerateTopic} disabled={isRecording}>
+            {isPending ? <Spinner /> : null}
             Generate Topic
           </Button>
           {currentTopic && (
