+++ package.json
@@ -8,6 +8,9 @@
     "start": "next start",
     "lint": "next lint"
   },
+  "prisma": {
+    "schema": "app/server/db/schema.prisma"
+  },
   "dependencies": {
     "@clerk/nextjs": "^6.9.5",
     "@prisma/client": "6.1.0",
